student-performance-backend/requirements.txt
Django
gunicorn
whitenoise
djangorestframework
django-cors-headers
scikit-learn
pandas
numpy
joblib
